requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pyjwt==2.8.0  
djangorestframework-simplejwt
channels==4.0.0
requests==2.31.0
gunicorn==21.2.0